fix(lwjgl3): fix lgtm issues

+++ engine/src/main/java/org/terasology/engine/subsystem/lwjgl/Lwjgl2Sync.java
@@ -143,7 +143,7 @@ public final class Lwjgl2Sync {
         public void dampenForLowResTicker() {
             if (this.avg() > DAMPEN_THRESHOLD) {
                 for (int i = 0; i < this.slots.length; i++) {
-                    this.slots[i] *= DAMPEN_FACTOR;
+                    this.slots[i] *= (long) DAMPEN_FACTOR;
                 }
             }
         }
